+++ L2J_Mobius_9.2_ReturnOfTheQueenAnt_Ch2/java/org/l2jmobius/gameserver/model/actor/instance/PetInstance.java
@@ -231,7 +231,7 @@ public class PetInstance extends Summon
 			pet.setTitle(owner.getName());
 			if (data.isSynchLevel() && (pet.getLevel() != owner.getLevel()))
 			{
-				final byte availableLevel = (byte) Math.min(data.getMaxLevel(), owner.getLevel());
+				final int availableLevel = Math.min(data.getMaxLevel(), owner.getLevel());
 				pet.getStat().setLevel(availableLevel);
 				pet.getStat().setExp(pet.getStat().getExpForLevel(availableLevel));
 			}
@@ -248,7 +248,7 @@ public class PetInstance extends Summon
 	 */
 	public PetInstance(NpcTemplate template, PlayerInstance owner, ItemInstance control)
 	{
-		this(template, owner, control, (byte) (template.getDisplayId() == 12564 ? owner.getLevel() : template.getLevel()));
+		this(template, owner, control, template.getDisplayId() == 12564 ? owner.getLevel() : template.getLevel());
 	}
 	
 	/**
@@ -258,13 +258,13 @@ public class PetInstance extends Summon
 	 * @param control
 	 * @param level
 	 */
-	public PetInstance(NpcTemplate template, PlayerInstance owner, ItemInstance control, byte level)
+	public PetInstance(NpcTemplate template, PlayerInstance owner, ItemInstance control, int level)
 	{
 		super(template, owner);
 		setInstanceType(InstanceType.PetInstance);
 		
 		_controlObjectId = control.getObjectId();
-		getStat().setLevel((byte) Math.max(level, PetDataTable.getInstance().getPetMinLevel(template.getId())));
+		getStat().setLevel(Math.max(level, PetDataTable.getInstance().getPetMinLevel(template.getId())));
 		_inventory = new PetInventory(this);
 		_inventory.restore();
