+++ L2J_Mobius_Essence_5.2_FrostLord/java/org/l2jmobius/gameserver/model/actor/stat/StaticObjectStat.java
@@ -34,6 +34,6 @@ public class StaticObjectStat extends CreatureStat
 	@Override
 	public int getLevel()
 	{
-		return (byte) getActiveChar().getLevel();
+		return getActiveChar().getLevel();
 	}
 }
